--- packages/playground/src/components/filters/constants/sort.ts
@@ -1,9 +1,0 @@
-import type { SortDirection } from '../types'
-
-/**
- * Available sort directions
- */
-export const SORT_DIRECTIONS: SortDirection[] = [
-  { label: 'Ascending', value: 'asc' },
-  { label: 'Descending', value: 'desc' }
-]
