+++ packages/playground/src/components/filters/index.ts
@@ -1,4 +1,9 @@
 import Filters from './filters'
 import FiltersSelectedBar from './filters-selected-bar'
+import useFilters from './use-filters'
 
-export { Filters, FiltersSelectedBar }
+import { FilterCondition, FilterOption, SortDirection, SortOption } from './types'
+
+export { Filters, FiltersSelectedBar, useFilters }
+
+export type { FilterOption, FilterCondition, SortOption, SortDirection }
