+++ packages/playground/src/pages/sandbox-repo-list-page.tsx
@@ -19,41 +19,238 @@ import { Link } from 'react-router-dom'
 import { mockRepos } from '../data/mockReposData'
 import { SandboxLayout } from '../index'
 import { PlaygroundSandboxLayoutSettings } from '../settings/sandbox-settings'
-import { Filters, FiltersSelectedBar } from '../components/filters'
-import useFilters from '../components/filters/use-filters'
+import {
+  Filters,
+  FiltersSelectedBar,
+  type FilterCondition,
+  type FilterOption,
+  type SortDirection,
+  type SortOption
+} from '../components/filters'
 
-import { getFilteredRepos, getFormattedReposWithDate, getSortedRepos } from '../components/filters/utils/repository'
-import { FILTER_OPTIONS, SORT_OPTIONS } from '../components/filters/constants/filter-and-sort-repo'
-import { SORT_DIRECTIONS } from '../components/filters/constants/sort'
+import useFilters from '../components/filters/use-filters'
+import { formatDistanceToNow } from 'date-fns'
+
+export const BASIC_CONDITIONS: FilterCondition[] = [
+  { label: 'is', value: 'is' },
+  { label: 'is not', value: 'is_not' },
+  { label: 'is empty', value: 'is_empty' },
+  { label: 'is not empty', value: 'is_not_empty' }
+]
+
+export const RANGE_CONDITIONS: FilterCondition[] = [
+  { label: 'is', value: 'is' },
+  { label: 'is before', value: 'is_before' },
+  { label: 'is after', value: 'is_after' },
+  { label: 'is between', value: 'is_between' },
+  { label: 'is empty', value: 'is_empty' },
+  { label: 'is not empty', value: 'is_not_empty' }
+]
+
+const FILTER_OPTIONS: FilterOption[] = [
+  {
+    label: 'Type',
+    value: 'type',
+    type: 'checkbox',
+    conditions: BASIC_CONDITIONS,
+    options: [
+      { label: 'Public', value: 'public' },
+      { label: 'Private', value: 'private' },
+      { label: 'Fork', value: 'fork' }
+    ]
+  },
+  {
+    label: 'Created time',
+    value: 'created_time',
+    type: 'date',
+    conditions: RANGE_CONDITIONS
+  }
+]
+
+const SORT_OPTIONS: SortOption[] = [
+  { label: 'Last updated', value: 'updated' },
+  { label: 'Stars', value: 'stars' },
+  { label: 'Forks', value: 'forks' },
+  { label: 'Pull Requests', value: 'pulls' },
+  { label: 'Title', value: 'title' }
+]
+
+const SORT_DIRECTIONS: SortDirection[] = [
+  { label: 'Ascending', value: 'asc' },
+  { label: 'Descending', value: 'desc' }
+]
 
 function SandboxRepoListPage() {
   const [loadState, setLoadState] = useState('float')
 
-  const {
-    activeFilters,
-    activeSorts,
-    handleFilterChange,
-    handleUpdateFilter,
-    handleUpdateCondition,
-    handleRemoveFilter,
-    handleSortChange,
-    handleUpdateSort,
-    handleRemoveSort,
-    handleResetFilters,
-    handleResetSorts,
-    handleReorderSorts,
-    handleResetAll,
-    searchQueries,
-    handleSearchChange
-  } = useFilters()
-
-  const filteredRepos = getFilteredRepos(mockRepos, activeFilters)
-  const sortedRepos = getSortedRepos(filteredRepos, activeSorts)
-  const reposWithFormattedDates = getFormattedReposWithDate(sortedRepos)
-
   const LinkComponent = ({ to, children }: { to: string; children: React.ReactNode }) => (
     <Link to={`/repos/${to}`}>{children}</Link>
   )
+
+  const filterHandlers = useFilters()
+
+  /**
+   * Filters repositories based on active filters
+   *
+   * @param repo - Repository object to filter
+   * @returns boolean - Whether the repository matches all active filters
+   *
+   * Filter logic:
+   * - If no active filters, returns all repositories
+   * - Applies all active filters (AND condition)
+   * - For type filter:
+   *   - Handles 'private', 'public', and 'fork' types
+   *   - Supports 'is' and 'is not' conditions
+   *   - Ignores empty filters or is_empty/is_not_empty conditions
+   */
+  const filteredRepos = mockRepos.filter(repo => {
+    if (filterHandlers.activeFilters.length === 0) return true
+
+    return filterHandlers.activeFilters.every(filter => {
+      switch (filter.type) {
+        case 'type': {
+          // Skip empty/not empty conditions for type filter
+          if (filter.condition === 'is_empty' || filter.condition === 'is_not_empty') {
+            return true
+          }
+
+          // Skip if no values selected
+          if (filter.selectedValues.length === 0) {
+            return true
+          }
+
+          // Determine repository types
+          const isPrivate = repo.private
+          const isPublic = !repo.private
+          const isFork = repo.forks > 0
+
+          // Check if repo matches any of the selected type values
+          const matchesType = filter.selectedValues.some(value => {
+            switch (value) {
+              case 'private':
+                return isPrivate
+              case 'public':
+                return isPublic
+              case 'fork':
+                return isFork
+              default:
+                return false
+            }
+          })
+
+          // Apply condition (is/is not)
+          return filter.condition === 'is' ? matchesType : !matchesType
+        }
+
+        case 'created_time': {
+          // Skip if no values selected
+          if (filter.selectedValues.length === 0) {
+            return true
+          }
+
+          // Handle empty conditions
+          if (filter.condition === 'is_empty') {
+            return !repo.createdAt
+          }
+          if (filter.condition === 'is_not_empty') {
+            return !!repo.createdAt
+          }
+
+          const createdDate = new Date(repo.createdAt)
+          const selectedDate = new Date(filter.selectedValues[0])
+
+          // Reset time parts for date-only comparison
+          createdDate.setHours(0, 0, 0, 0)
+          selectedDate.setHours(0, 0, 0, 0)
+
+          switch (filter.condition) {
+            case 'is':
+              return createdDate.getTime() === selectedDate.getTime()
+
+            case 'is_before':
+              return createdDate.getTime() < selectedDate.getTime()
+
+            case 'is_after':
+              return createdDate.getTime() > selectedDate.getTime()
+
+            case 'is_between': {
+              if (filter.selectedValues.length !== 2) return true
+              const endDate = new Date(filter.selectedValues[1])
+              endDate.setHours(0, 0, 0, 0)
+              return createdDate.getTime() >= selectedDate.getTime() && createdDate.getTime() <= endDate.getTime()
+            }
+
+            default:
+              return true
+          }
+        }
+
+        default:
+          return true
+      }
+    })
+  })
+
+  /**
+   * Sorts filtered repositories based on active sorts
+   *
+   * @param a - First repository to compare
+   * @param b - Second repository to compare
+   * @returns number - Comparison result (-1, 0, 1)
+   *
+   * Sort logic:
+   * - Applies multiple sorts in priority order (first sort has highest priority)
+   * - For equal values, moves to next sort criteria
+   * - Supports following sort types:
+   *   - updated: by timestamp
+   *   - stars: by number of stars
+   *   - forks: by number of forks
+   *   - pulls: by number of pull requests
+   *   - title: alphabetically by name
+   * - Each sort supports ascending/descending direction
+   */
+  const sortedRepos = [...filteredRepos].sort((a, b) => {
+    // Iterate through sorts in priority order
+    for (const sort of filterHandlers.activeSorts) {
+      const direction = sort.direction === 'asc' ? 1 : -1
+      let comparison = 0
+
+      switch (sort.type) {
+        case 'updated':
+          comparison = (new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()) * direction
+          break
+        case 'stars':
+          comparison = (b.stars - a.stars) * direction
+          break
+        case 'forks':
+          comparison = (b.forks - a.forks) * direction
+          break
+        case 'pulls':
+          comparison = (b.pulls - a.pulls) * direction
+          break
+        case 'title':
+          comparison = a.name.localeCompare(b.name) * direction
+          break
+        default:
+          comparison = 0
+      }
+
+      // If items are different by current sort criteria, return result
+      if (comparison !== 0) {
+        return comparison
+      }
+    }
+
+    return 0
+  })
+
+  const reposWithFormattedDates = sortedRepos.map(repo => ({
+    ...repo,
+    timestamp: formatDistanceToNow(new Date(repo.createdAt), {
+      addSuffix: true,
+      includeSeconds: true
+    }).replace('about ', '')
+  }))
 
   return (
     <>
@@ -89,39 +286,16 @@ function SandboxRepoListPage() {
               <SearchBox.Root placeholder="Search repositories" />
             </ListActions.Left>
             <ListActions.Right>
-              <Filters
-                filterOptions={FILTER_OPTIONS}
-                sortOptions={SORT_OPTIONS}
-                activeFilters={activeFilters}
-                activeSorts={activeSorts}
-                onFilterChange={handleFilterChange}
-                onSortChange={handleSortChange}
-                onResetFilters={handleResetFilters}
-                onResetSort={handleResetSorts}
-                searchQueries={searchQueries}
-                onSearchChange={handleSearchChange}
-              />
+              <Filters filterOptions={FILTER_OPTIONS} sortOptions={SORT_OPTIONS} {...filterHandlers} />
               <Button variant="default">New repository</Button>
             </ListActions.Right>
           </ListActions.Root>
-          {(activeFilters.length > 0 || activeSorts.length > 0) && <Spacer size={2} />}
+          {(filterHandlers.activeFilters.length > 0 || filterHandlers.activeSorts.length > 0) && <Spacer size={2} />}
           <FiltersSelectedBar
-            activeFilters={activeFilters}
-            activeSorts={activeSorts}
             filterOptions={FILTER_OPTIONS}
             sortOptions={SORT_OPTIONS}
             sortDirections={SORT_DIRECTIONS}
-            onRemoveFilter={handleRemoveFilter}
-            onUpdateFilter={handleUpdateFilter}
-            onUpdateCondition={handleUpdateCondition}
-            onUpdateSort={handleUpdateSort}
-            onRemoveSort={handleRemoveSort}
-            onSortChange={handleSortChange}
-            onResetSorts={handleResetSorts}
-            onReorderSorts={handleReorderSorts}
-            onResetAll={handleResetAll}
-            searchQueries={searchQueries}
-            onSearchChange={handleSearchChange}
+            {...filterHandlers}
           />
           <Spacer size={5} />
           <RepoList repos={reposWithFormattedDates} LinkComponent={LinkComponent} />
